filter menu displays only categories from that restaurant
