+++ frontend/src/App.tsx
@@ -10,7 +10,7 @@ import {
 import {AppProvider} from "./provider/app.provider.tsx";
 import {MainLayout} from "./layouts/main.layout.tsx";
 import {ProfilePage} from "./pages/profile.page.tsx"
-import { RestaurantPage } from "./pages/restaurant-page.tsx";
+import { RestaurantPage } from "./pages/restaurant.page.tsx";
 
 //Todo: define routes in separate file
 export default function App() {
